move CommandResult
